[refactor] add zones package to hold all cell-related code
