+++ src/graph/main.ts
@@ -1,8 +1,8 @@
 import {Feature, FeatureCollection, featureCollection, LineString, Polygon} from "@turf/helpers";
 import {Shape} from "../Shape";
 import envelope from "@turf/envelope";
-import {getCellsFromArea} from "../getCellsFromArea";
-import {markCellsWithPaths} from "../markCellsWithPaths";
+import {getCellsFromArea} from "../zones/getCellsFromArea";
+import {markCellsWithPaths} from "../zones/markCellsWithPaths";
 import {convertPolygonsToGraph} from "./convertPolygonsToGraph";
 import Graph from "graphology";
