+++ src/zones/colorCells.ts
