+++ src/zones/markCellsWithPaths.ts
